doc algorithm tweak
As in its example, the last line is not as indented as the contents (a relatively normal practice any other code besides raw strings).

In the original algorithm, this would have left a level of indentation (4 spaces, here) in each line. The adjustment corrects for this, by handling the last line differently.

I've also added a note on empty lines. If your editor (or git) strips trailing spaces, this way it will not affect the result.